refactor(sense): split Effects stories into Confetti, LogoLoop, and Icons
One story file per component instead of a combined "Effects & icons" entry;
icons get an individual story each (Check, Megaphone).
